Add files via upload
This is the main file for the TimeSeries data model. It begins with loading the data, then plotting and visualizing the data for an increased understanding of what is being worked on. Then the model creation begins, starting with the initial TimeSeries split, then using a test set of data to forecast initial predictions. After saving the model, the final section of code is a Bayesian optimization procedure designed to hypertune the model's parameters.